+++ lib/events_api_web/views/user_view.ex
@@ -16,7 +16,6 @@ defmodule EventsApiWeb.UserView do
     }
   end
 
-
   def render("user.json", %{user: user}) do
     %{
       id: user.id,